Switched sail commands to Float32 for intermediate-position measurement

+++ fit_pc_pkg/scripts/RC-control-master.py
@@ -26,7 +26,7 @@ class RCmaster(object):
         rospy.init_node('command_center', anonymous=True)
         
         self.rudderPub = rospy.Publisher('rc_mode/rudder/set_point', Int16)
-        self.sailPub = rospy.Publisher('rc_mode/sail/set_point', Int16)
+        self.sailPub = rospy.Publisher('rc_mode/sail/set_point', Float32)
 
         self.sailSub = rospy.Subscriber('rc/sails_in', Float32, self.update_sail)
         self.rudderSub = rospy.Subscriber('rc/rudder_in', Float32, self.update_rudder)
